+++ hypopg_index.c
@@ -8,7 +8,7 @@
  * This program is open source, licensed under the PostgreSQL license.
  * For license terms, see the LICENSE file.
  *
- * Copyright (C) 2015-2018: Julien Rouhaud
+ * Copyright (C) 2015-2021: Julien Rouhaud
  *
  *-------------------------------------------------------------------------
  */
@@ -57,6 +57,9 @@
 #endif
 #include "parser/parse_utilcmd.h"
 #include "parser/parser.h"
+#if PG_VERSION_NUM >= 120000
+#include "port/pg_bitutils.h"
+#endif
 #include "storage/bufmgr.h"
 #include "utils/builtins.h"
 #include "utils/lsyscache.h"
@@ -173,6 +176,7 @@ hypo_newIndex(Oid relid, char *accessMethod, int nkeycolumns, int ninccolumns,
 	entry->amcanorder = amroutine->amcanorder;
 #if PG_VERSION_NUM >= 110000
 	entry->amcanparallel = amroutine->amcanparallel;
+	entry->amcaninclude = amroutine->amcaninclude;
 #endif
 #else
 	/* Up to 9.5, all information is available in the pg_am tuple */
@@ -251,6 +255,14 @@ hypo_newIndex(Oid relid, char *accessMethod, int nkeycolumns, int ninccolumns,
 #if PG_VERSION_NUM >= 90600
 			&& entry->relam != BLOOM_AM_OID
 #endif
+#if PG_VERSION_NUM >= 100000
+			/*
+			 * Only support hash indexes for pg10+.  In previous version they
+			 * weren't crash safe, and changes in pg10+ also significantly
+			 * changed the disk space allocation.
+			 */
+			 && entry->relam != HASH_AM_OID
+#endif
 			)
 		{
 			/*
@@ -311,6 +323,9 @@ hypo_index_reset(void)
 
 	list_free(hypoIndexes);
 	hypoIndexes = NIL;
+
+	hypo_reset_fake_oids();
+
 	return;
 }
 
@@ -439,6 +454,14 @@ hypo_index_store_parsetree(IndexStmt *node, const char *queryString)
 					(errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
 					 errmsg("hypopg: access method \"%s\" does not support multicolumn indexes",
 							node->accessMethod)));
+
+#if PG_VERSION_NUM >= 110000
+		if (node-> indexIncludingParams != NIL && !entry->amcaninclude)
+			ereport(ERROR,
+					(errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
+					 errmsg("hypopg: access method \"%s\" does not support included columns",
+							node->accessMethod)));
+#endif
 
 		entry->unique = node->unique;
 		entry->ncolumns = nkeycolumns + ninccolumns;
@@ -1116,35 +1139,40 @@ hypo_injectHypotheticalIndex(PlannerInfo *root,
 	rel->indexlist = lcons(index, rel->indexlist);
 }
 
-/* Return the hypothetical index name is indexId is ours, NULL otherwise, as
+/*
+ * Return the stored hypothetical index for a given oid if any, NULL otherwise
+ */
+hypoIndex *
+hypo_get_index(Oid indexId)
+{
+	ListCell   *lc;
+
+	foreach(lc, hypoIndexes)
+	{
+		hypoIndex  *entry = (hypoIndex *) lfirst(lc);
+
+		if (entry->oid == indexId)
+			return entry;
+	}
+
+	return NULL;
+}
+
+/* Return the hypothetical index name ifs indexId is ours, NULL otherwise, as
  * this is what explain_get_index_name expects to continue his job.
  */
 const char *
 hypo_explain_get_index_name_hook(Oid indexId)
 {
-	char	   *ret = NULL;
-
 	if (isExplain)
 	{
-		/*
-		 * we're in an explain-only command. Return the name of the
-		 * hypothetical index name if it's one of ours, otherwise return NULL
-		 */
-		ListCell   *lc;
+		hypoIndex  *index = NULL;
 
-		foreach(lc, hypoIndexes)
-		{
-			hypoIndex  *entry = (hypoIndex *) lfirst(lc);
+		index = hypo_get_index(indexId);
 
-			if (entry->oid == indexId)
-			{
-				ret = entry->indexname;
-			}
-		}
+		if (index)
+			return index->indexname;
 	}
-
-	if (ret)
-		return ret;
 
 	if (prev_explain_get_index_name_hook)
 		return prev_explain_get_index_name_hook(indexId);
@@ -1838,6 +1866,110 @@ hypo_estimate_index(hypoIndex * entry, RelOptInfo *rel)
 		entry->pages = 1;		/* meta page */
 		entry->pages += (BlockNumber) ceil(
 										   ((double) entry->tuples * line_size) / usable_page_size);
+	}
+#endif
+#if PG_VERSION_NUM >= 100000
+	else if (entry->relam == HASH_AM_OID)
+	{
+		/* ----------------------------
+		 * From hash AM readme (src/backend/access/hash/README):
+		 *
+		 *   There are four kinds of pages in a hash index: the meta page (page
+		 *   zero), which contains statically allocated control information;
+		 *   primary bucket pages; overflow pages; and bitmap pages, which keep
+		 *   track of overflow pages that have been freed and are available for
+		 *   re-use.  For addressing purposes, bitmap pages are regarded as a
+		 *   subset of the overflow pages.
+		 * [...]
+		 *   A hash index consists of two or more "buckets", into which tuples
+		 *   are placed whenever their hash key maps to the bucket number.
+		 *   [...]
+		 *   Each bucket in the hash index comprises one or more index pages.
+		 *   The bucket's first page is permanently assigned to it when the
+		 *   bucket is created.  Additional pages, called "overflow pages", are
+		 *   added if the bucket receives too many tuples to fit in the primary
+		 *   bucket page.
+		 *
+		 * Hash AM also already provides some functions to compute an initial
+		 * number of buckets given the estimated number of tuples the index
+		 * will contains, which is a good enough estimate for hypothetical
+		 * index.
+		 *
+		 * The code below is simply an adaptation of original code to compute
+		 * the initial number of bucket, modified to cope with hypothetical
+		 * index, plus some naive estimates for the overflow and bitmap pages.
+		 *
+		 * For more details, refer to the original code, in:
+		 *   - _hash_init()
+		 *   - _hash_init_metabuffer()
+		 */
+		int32 data_width;
+		int32 item_width;
+		int32 ffactor;
+		double dnumbuckets;
+		uint32 num_buckets;
+		uint32 num_overflow;
+		uint32 num_bitmap;
+		uint32 lshift;
+
+	/*
+	 * Determine the target fill factor (in tuples per bucket) for this index.
+	 * The idea is to make the fill factor correspond to pages about as full
+	 * as the user-settable fillfactor parameter says.  We can compute it
+	 * exactly since the index datatype (i.e. uint32 hash key) is fixed-width.
+	 */
+	data_width = sizeof(uint32);
+	item_width = MAXALIGN(sizeof(IndexTupleData)) + MAXALIGN(data_width) +
+		sizeof(ItemIdData);		/* include the line pointer */
+	ffactor = HypoHashGetTargetPageUsage(fillfactor) / item_width;
+	/* keep to a sane range */
+	if (ffactor < 10)
+		ffactor = 10;
+
+	/*
+	 * Choose the number of initial bucket pages to match the fill factor
+	 * given the estimated number of tuples.  We round up the result to the
+	 * total number of buckets which has to be allocated before using its
+	 * hashm_spares element. However always force at least 2 bucket pages. The
+	 * upper limit is determined by considerations explained in
+	 * _hash_expandtable().
+	 */
+	dnumbuckets = entry->tuples / ffactor;
+	if (dnumbuckets <= 2.0)
+		num_buckets = 2;
+	else if (dnumbuckets >= (double) 0x40000000)
+		num_buckets = 0x40000000;
+	else
+		num_buckets = _hash_get_totalbuckets(_hash_spareindex(dnumbuckets));
+
+	/*
+	 * Naive estimate of overflow pages, knowing that a page can store ffactor
+	 * tuples: we compute the number of tuples that wouldn't fit in the
+	 * previously computed number of buckets, and compute the number of pages
+	 * needed to store them.
+	 */
+	num_overflow = Max(0, ((entry->tuples - (num_buckets * ffactor)) /
+					   ffactor) + 1);
+
+	/* find largest bitmap array size that will fit in page size */
+#if PG_VERSION_NUM >= 120000
+	lshift = pg_leftmost_one_pos32(HypoHashGetMaxBitmapSize());
+#else
+	for (lshift = _hash_log2(HypoHashGetMaxBitmapSize()); lshift > 0; --lshift)
+	{
+		if ((1 << lshift) <= HypoHashGetMaxBitmapSize())
+			break;
+	}
+#endif
+
+	/*
+	 * Naive estimate of bitmap pages, using the previously computed number of
+	 * overflow pages.
+	 */
+	num_bitmap = Max(1, num_overflow / (1 <<lshift));
+
+	/* Simply add all computed pages, plus one extra block for the meta page */
+	entry->pages = num_buckets + num_overflow + num_bitmap + 1;
 	}
 #endif
 	else